Change the way dependencies are resolved
The requiredLibraries and requiredIncludes properties are problematic if
they cause binaries to be resolved prematurely. They are being
deprecated in favor of a new closure, withResolvedLibraries. This
closure appears inside buildConfig and allows safely accessing the
resolved lists of required libraries and headers.

Also added a lib() helper inside buildConfig, so libraries can be more
easily included along with the rest of the config.