Refactor task creation again
Version 1.8 turned out to be pretty broken, so this is another attempt.
It uses @BinaryTasks to create the external build tasks, and then a
combination of @Finalize and @Each to configure them. This seems to
allow the configuration to work using the original syntax, without
circular dependency errors from resolving binary.getLibs().