Allow multiple buildConfig/buildOutput blocks
The specs will now track a list of config/output actions, instead of a
single action. This means that both buildConfig and buildOutput can be
specified an arbitrary number of times for a given component, and they
will be aggregated into a single composite action. The task type,
however, can only be set once.